requirements.txt
Django
Pillow
pytz

django-flat-theme
django-sitetree
django-compressor
django-debug-toolbar
django-ckeditor
django-model-utils
